LPS-75792 Move MonitoringConfiguration into internal package
