+++ modules/apps/foundation/portal/portal-monitoring/src/main/java/com/liferay/portal/monitoring/internal/configuration/MonitoringConfiguration.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.monitoring.configuration;
+package com.liferay.portal.monitoring.internal.configuration;
 
 import aQute.bnd.annotation.metatype.Meta;
